Add closing bracket to exclude list in regex
If a Cover block is used the image is in an inline style e.g.
style="background-image:url(https://example.com/wp-content/uploads/2020/08/image.jpg)"
The regex was capturing the ) in the match i.e. https://example.com/wp-content/uploads/2020/08/image.jpg)
and this was breaking the site when the image file exists on the staging (non-production) site.